LayoutTests/canvas/philip/tests/2d.path.isPointInPath.winding-expected.txt

< [index] >
2d.path.isPointInPath.winding
isPointInPath() uses the non-zero winding number rule
References: 2d.path.isPointInPath
Actual output:
Passed
